Training data of Titanic final voyage passengers
From Kaggle Titanic Challenge.  This is the training (training.csv) dataset used in the challenge.  This dataset includes attributes for passengers on the final (doomed) voyage of the White Star Line ship, the Titanic.  For predictive modeling, the 'survived' attribute is typically the target variable.  While the original file is named 'training.csv' this copy of that file is named 'titanic.csv' to avoid ambiguity.